Add builder to CfDnsRecords and use same collection everywhere
